Fixes size_t warnings, fixes wrong order warnings, removes unused vars

+++ APGG/Organism.cpp
@@ -67,7 +67,7 @@ namespace APGG {
         copyOrganism->m_mutated = m_mutated;
 		copyOrganism->m_children = m_children;
 		copyOrganism->m_parent = m_parent;
-		for (int i = 0; i < m_genomes.size(); i++) {
+		for (size_t i = 0; i < m_genomes.size(); i++) {
 			copyOrganism->m_genomes[i].setValue(m_genomes[i].getValue());
 		}
 	}
